fix: resolve property filtering
